fix: move mic button outside scroll area so it sticks to bottom of Moments tab
Renders the button as a flex sibling to the scrollable tab content in
AdminPanelV4, so it stays fixed at the bottom regardless of scroll position.
